requirements.txt
aiohttp==3.6.2
aiosqlite==0.11.0
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
chardet==3.0.4
ciso8601==2.1.3
Click==7.0
decorator==4.4.1
h11==0.9.0
httptools==0.0.13
idna==2.8
multidict==4.7.4
pip==9.0.1
pkg-resources==0.0.0
PyPika==0.35.21
python-dotenv==0.11.0
setuptools==39.0.1
six==1.14.0
soupsieve==1.9.5
starlette==0.13.1
tortoise-orm==0.15.9
typing-extensions==3.7.4.1
uvicorn==0.11.2
uvloop==0.14.0
validators==0.14.2
websockets==8.1
yarl==1.4.2
